requirement.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
openai
pyyaml
joblib
